requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.14
    # via semantic-kernel
aioice==0.10.1
    # via aiortc
aiortc==1.13.0
    # via semantic-kernel
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
av==14.4.0
    # via aiortc
azure-ai-agents==1.0.2
    # via
    #   azure-ai-projects
    #   semantic-kernel
azure-ai-inference==1.0.0b9
    # via agent-hub-python (pyproject.toml)
azure-ai-projects==1.0.0b12
    # via
    #   agent-hub-python (pyproject.toml)
    #   semantic-kernel
azure-core==1.35.0
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-identity
    #   azure-storage-blob
azure-identity==1.23.1
    # via
    #   agent-hub-python (pyproject.toml)
    #   semantic-kernel
azure-storage-blob==12.26.0
    # via
    #   agent-hub-python (pyproject.toml)
    #   azure-ai-projects
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via
    #   aiortc
    #   cryptography
    #   pylibsrtp
chardet==5.2.0
    # via prance
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via uvicorn
cloudevents==1.12.0
    # via semantic-kernel
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==45.0.5
    # via
    #   aiortc
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
    #   pyopenssl
defusedxml==0.7.1
    # via semantic-kernel
deprecation==2.1.0
    # via cloudevents
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via aioice
fastapi==0.116.1
    # via agent-hub-python (pyproject.toml)
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
google-crc32c==1.7.1
    # via aiortc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
ifaddr==0.2.0
    # via aioice
importlib-metadata==8.7.0
    # via opentelemetry-api
isodate==0.7.2
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-storage-blob
    #   openapi-core
jinja2==3.1.6
    # via semantic-kernel
jiter==0.10.0
    # via openai
jsonschema==4.25.0
    # via
    #   openapi-core
    #   openapi-schema-validator
    #   openapi-spec-validator
jsonschema-path==0.3.4
    # via
    #   openapi-core
    #   openapi-spec-validator
jsonschema-specifications==2025.4.1
    # via
    #   jsonschema
    #   openapi-schema-validator
lazy-object-proxy==1.11.0
    # via openapi-spec-validator
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
more-itertools==10.7.0
    # via openapi-core
msal==1.32.3
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
multidict==6.6.3
    # via
    #   aiohttp
    #   yarl
nest-asyncio==1.6.0
    # via semantic-kernel
numpy==2.3.1
    # via
    #   scipy
    #   semantic-kernel
openai==1.97.0
    # via
    #   agent-hub-python (pyproject.toml)
    #   semantic-kernel
openapi-core==0.19.5
    # via semantic-kernel
openapi-schema-validator==0.6.3
    # via
    #   openapi-core
    #   openapi-spec-validator
openapi-spec-validator==0.7.2
    # via openapi-core
opentelemetry-api==1.35.0
    # via
    #   agent-hub-python (pyproject.toml)
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   semantic-kernel
opentelemetry-sdk==1.35.0
    # via semantic-kernel
opentelemetry-semantic-conventions==0.56b0
    # via opentelemetry-sdk
packaging==25.0
    # via
    #   deprecation
    #   prance
parse==1.20.2
    # via openapi-core
pathable==0.4.4
    # via jsonschema-path
prance==25.4.8.0
    # via semantic-kernel
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
protobuf==6.31.1
    # via semantic-kernel
pybars4==0.9.13
    # via semantic-kernel
pycparser==2.22
    # via cffi
pydantic==2.11.7
    # via
    #   fastapi
    #   openai
    #   pydantic-settings
    #   semantic-kernel
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via semantic-kernel
pyee==13.0.0
    # via aiortc
pyjwt==2.10.1
    # via msal
pylibsrtp==0.12.0
    # via aiortc
pymeta3==0.5.1
    # via pybars4
pyopenssl==25.1.0
    # via aiortc
python-dotenv==1.1.1
    # via
    #   agent-hub-python (pyproject.toml)
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.2
    # via
    #   jsonschema-path
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-path
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   azure-core
    #   jsonschema-path
    #   msal
    #   prance
rfc3339-validator==0.1.4
    # via openapi-schema-validator
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.14
    # via prance
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
scipy==1.16.0
    # via semantic-kernel
semantic-kernel==1.32.2
    # via agent-hub-python (pyproject.toml)
six==1.17.0
    # via
    #   azure-core
    #   rfc3339-validator
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.47.1
    # via fastapi
tqdm==4.67.1
    # via openai
typing-extensions==4.14.1
    # via
    #   azure-ai-agents
    #   azure-ai-inference
    #   azure-ai-projects
    #   azure-core
    #   azure-identity
    #   azure-storage-blob
    #   fastapi
    #   openai
    #   openapi-core
    #   opentelemetry-api
    #   opentelemetry-sdk
    #   opentelemetry-semantic-conventions
    #   pydantic
    #   pydantic-core
    #   pyee
    #   semantic-kernel
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn==0.35.0
    # via agent-hub-python (pyproject.toml)
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via
    #   semantic-kernel
    #   uvicorn
werkzeug==3.1.1
    # via openapi-core
yarl==1.20.1
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
